Complete detector lib. All tests have passed. Start working on cloud api

+++ src/cloud/src/WebDavClientApi.cs
@@ -1,4 +1,15 @@
-﻿using System;
+﻿/***************************************************************************
+ *                                                                         *
+ *                                                                         *
+ * Copyright(c) 2019-2020, REGATA Experiment at FLNP|JINR                  *
+ * Author: [Boris Rumyantsev](mailto:bdrum@jinr.ru)                        *
+ *                                                                         *
+ * The REGATA Experiment team license this file to you under the           *
+ * GNU GENERAL PUBLIC LICENSE                                              *
+ *                                                                         *
+ ***************************************************************************/
+
+using System;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,7 +20,7 @@ using System.Threading;
 using System.Xml.Linq;
 using System.Collections.Generic;
 
-namespace Extensions
+namespace Regata.Core.Cloud
 {
     public static class WebDavClientApi
     {
@@ -17,27 +28,38 @@ namespace Extensions
         private const string _hostBase = @"https://disk.jinr.ru";
         private const string _hostWebDavAPI = @"/remote.php/dav/files/regata";
         private const string _hostOCSApi = @"/ocs/v2.php/apps/files_sharing/api/v1/shares";
-        public const string DiskJinrTarget = "MeasurementsDiskJinr";
+        private static string _diskJinrTarget;
+        public static string GetDownloadLink(string sharedKey) => $"https://disk.jinr.ru/index.php/s/{sharedKey}/download";
+
+        public static string DiskJinrTarget
+        {
+            get { return _diskJinrTarget; }
+
+            set
+            {
+                _diskJinrTarget = value;
+                var cm = AdysTech.CredentialManager.CredentialManager.GetCredentials(DiskJinrTarget);
+
+                if (cm == null)
+                    throw new ArgumentException("Can't load cloud storage credential. Please add it to the windows credential manager");
+
+                _httpClient = new HttpClient();
+                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Authorization", "Basic " + Convert.ToBase64String(Encoding.UTF8.GetBytes($"{cm.UserName}:{cm.Password}")));
+            }
+        }
 
         public static void Cancel()
         {
             _httpClient.CancelPendingRequests();
         }
 
-        static WebDavClientApi()
-        {
-            var cm = AdysTech.CredentialManager.CredentialManager.GetCredentials(DiskJinrTarget);
 
-            if (cm == null)
-               throw new ArgumentException("Can't load cloud storage credential. Please add it to the windows credential manager");
-
-            _httpClient = new HttpClient();
-            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Authorization", "Basic " + Convert.ToBase64String(Encoding.UTF8.GetBytes($"{cm.UserName}:{cm.Password}")));
-        }
 
         public static async Task<bool> RemoveFile(string path, CancellationToken ct)
         {
-            if (! await IsExists(path, ct)) return true;
+            if (_httpClient == null) return false;
+
+            if (!await IsExists(path, ct)) return true;
             var response = await _httpClient.DeleteAsync($"{_hostBase}{_hostWebDavAPI}/{path.Substring(Path.GetPathRoot(path).Length)}", ct).ConfigureAwait(false);
 
             return IsSuccessfull(await response.Content.ReadAsStringAsync());
@@ -45,6 +67,8 @@ namespace Extensions
 
         public static async Task<bool> UploadFile(string path, CancellationToken ct)
         {
+            if (_httpClient == null) return false;
+
             if (!File.Exists(path)) throw new FileNotFoundException($"File '{path}' doesn't exist");
             await CreateFolder(Path.GetDirectoryName(path), ct);
             using (HttpContent bytesContent = new ByteArrayContent(File.ReadAllBytes(path)))
@@ -56,6 +80,8 @@ namespace Extensions
 
         public static async Task<string> MakeShareable(string file, CancellationToken ct)
         {
+            if (_httpClient == null) return string.Empty;
+
             using (var request = new HttpRequestMessage(new HttpMethod("POST"), $"{_hostBase}{_hostOCSApi}?path={file.Substring(Path.GetPathRoot(file).Length)}&shareType=3&permissions=3&name={Path.GetFileNameWithoutExtension(file)}"))
             {
                 request.Headers.Add("OCS-APIRequest", "true");
@@ -73,12 +99,17 @@ namespace Extensions
 
         public static async Task<bool> IsExists(string path, CancellationToken ct)
         {
+            if (_httpClient == null) return false;
+
             var response = await Send(new Uri($"{_hostBase}{_hostWebDavAPI}/{path.Substring(Path.GetPathRoot(path).Length)}"), new HttpMethod("PROPFIND"), ct);
             return IsSuccessfull(await response.Content.ReadAsStringAsync());
         }
 
         public static async Task CreateFolder(string path, CancellationToken ct)
         {
+            if (_httpClient == null) return;
+
+
             var dir = path.Substring(Path.GetPathRoot(path).Length);
             var subPath = "";
             foreach (var node in dir.Split(Path.DirectorySeparatorChar))
@@ -94,6 +125,9 @@ namespace Extensions
 
         private static bool IsSuccessfull(string resp)
         {
+
+            if (_httpClient == null) return false;
+
             if (resp.Contains("exception") || resp.Contains("error"))
                 return false;
             return true;
@@ -106,6 +140,9 @@ namespace Extensions
             KeyValuePair<string, string>? header = null,
             HttpCompletionOption httpCompletionOption = HttpCompletionOption.ResponseContentRead)
         {
+            if (_httpClient == null) return null;
+
+
             using (var request = new HttpRequestMessage(method, requestUri))
             {
                 if (header != null)
@@ -113,6 +150,22 @@ namespace Extensions
                 return await _httpClient.SendAsync(request, HttpCompletionOption.ResponseContentRead, cancellationToken).ConfigureAwait(false);
             }
         }
-    }
 
-}
+        public static async Task DownloadFile(string shareId, string path, CancellationToken ct)
+        {
+            if (!Directory.Exists(Path.GetDirectoryName(path)))
+                Directory.CreateDirectory(Path.GetDirectoryName(path));
+
+            using (var request = new HttpRequestMessage(HttpMethod.Get, GetDownloadLink(shareId)))
+            {
+                using (
+                    Stream contentStream = await (await _httpClient.SendAsync(request, ct)).Content.ReadAsStreamAsync(),
+                    stream = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None, 4096, true))
+                {
+                    await contentStream.CopyToAsync(stream, 4096, ct);
+                }
+            }
+        }
+
+    } // public static class WebDavClientApi    
+}     // namespace Regata.Core.Cloud
